moodsync/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
pandas==2.0.3
scikit-learn==1.3.0
python-dateutil==2.8.2
werkzeug==2.3.7